Fix: workaround for using tools with no go directive in their go.mod (#156)
* workaround for using tools with no ge directive in their go.mod
According to https://go.dev/ref/mod#go-mod-file-go, prior to Go 1.21,
the go directive was option.
This seems to not be supported in bingo. This commit only changes the fallback value used when a go.mod file does not contain the version. This semantically means that any go version should be able to compile this.

I encountered the issue while trying to use bingo to install github.com/betacraft/easytags
Their go.mod is completely empty, which causes a panic on

if semver.MustParse(targetModParsed.GoVersion()).GreaterThan(runnable.GoVersion()) {

`get.go:723`

* fix test and add doc comment

---------

Co-authored-by: dror nir <me@drornir.dev>